Expose kernel tracer to user-space (version 5)
By writing to the file /proc/lttng, a user-space application creates a
kernel event. The event's payload is by default UTF-8 text, but any data
can be written, up to 1024 bytes. Null-character is optional and is not
enforced. The event uses sequence for space efficiency and to store any
data as payload.

The feature is enabled when both lttng-uevent and lttng-probe-uevent are
loaded. The lttng-abi module exports a register function and includes a wrapper
for lttng_fops write. This is required since struct file_operations must be
const.

Module unload must be prevented while lttng-uevent is being used.
rcu_synchronized() is invoked on module unload, and thus rcu_read_lock() can be
used to force waiting until the critical section is over.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>