Expose kernel tracer to user-space
By writing to the file /proc/lttng, a user-space application creates a
kernel event. The event's payload is by default UTF-8 text, but any data
can be written, up to 1024 bytes. Null-character is optional and is not
enforced. The event uses sequence for space efficiency and to store any
data as payload.

Update: split the probe code into it's own module and make it an optional
feature of lttng-abi. The feature is enabled when the module is loaded. The
lttng-abi module exports a register function and includes a wrapper for
lttng_fops write. This is required since struct file_operations must be const.
Since the module dependency is reversed, unloading the lttng-uevent module is
done only when it's not used anymore. This is done with rwlock synchronisation.
The synchronisation doesn't prevent starvation, but this situation is unlikely
and can be prevented by stop active tracing sessions.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>