Expose kernel tracer to user-space
By writing to the file /proc/lttng, a user-space application creates a
kernel event. The event's payload is by default UTF-8 text, but any data
can be written, up to 1024 bytes. Null-character is optional and is not
enforced. The event uses sequence for space efficiency and to store any
data as payload.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>