Use reference to module
Instead of using rwlock to prevent removing the module lttng_uevent while it's
in use, get a reference to the module.

The problem with this approach is that the module can't be unloaded, because
lttng_tracer module has a reference to it, and rmmod refuse to remove it, such
that the function module_exit, that release the refcount, is never called.