Fix : hostname field in statedump process state
The hostname can be added as a context and is present in the environment
but is not in the statedump, so until a process has any activity we
don't know which hostname it is associated with.

This patch adds the hostname field in the statedump process state.

Note that in case of contextualisation (with containers), a process can
be in a hierarchy of PID namespaces but in only on UTS namespace (this
namespace is not hierarchical).

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>