Metadata cache and multiple metadata channels
Implement a cache for the kernel metadata and a list of metadata
channels.
When new metadata is appended, it is sent to all the metadata channels
enabled. When a new metadata channel is created, it receives all the
cache content.
This allows to request the metadata multiple times.

With this new feature, the poll and get_subbuf ring buffer operations
are now overridden by lttng-abi for the metadata channels, to check the
cache before doing these operations.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>