Add sched_ttwu event
The sched_wakeup event occurs on the target CPU on linux > 3.8.0. It prevent to
make the required link between the blocked task and the awakener.

To fix this problem, we register a kprobe on the function try_to_wake_up(),
that is always called for a wakeup, but before the IPI.

There are two drawbacks of this method. First, it makes mandatory to use
lttng-modules addons to make the critical path analysis. Also, there is a
greater overhead (~1us) per event recorded.

Make it work, then make it fast. Probably a new tracepoint in that location
could be added.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>