Add events required to vm synchronization
Register probe to soft_irq_exit instead of timer

When a VM is idle, its vCPU gets awaken by the timer. We want to disable
the hypercall in case of and idle vCPU. We can use the nature of
tickless kernel to follow the scheduling of interrupts, which is delayed
on an idle system. We use a homemade rate limit to avoid triggering a
hypercall on each softirq. We use a homemade way of doing all the hypercalls
on the same vCPU.

Use one counter per virtual machine.

Signed-off-by: Mohamad Gebai <mohamad.gebai@polymtl.ca>
Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>