Do not lazy load probes and name check
Addons modules are loaded after the main lttng modules. If the probes are lazy
loaded, then the addons event are not listed, unless a session is started
first. To make sure events are shown, immediately load the probes.

Also, remove the check for the event names. It works perfectly even though the
event name does not starts with the provider name. An event name clash can
occur, just don't do that.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>